Move helper functions to helpers folder
